Init Notification Table

+++ app/Helper/RazkyFeb.php
@@ -3,59 +3,66 @@
 namespace App\Helper;
 
 
+use App\Models\UserMNotification;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\File;
+use Illuminate\Support\Str;
 
 class RazkyFeb
 {
 
-    public static function hello(){
+    public static function hello()
+    {
         return "hello";
     }
 
     public static function isAPI()
     {
         $url = url()->current();
-        if (str_contains($url, 'api/')){
+        if (str_contains($url, 'api/')) {
             return true;
-        }else{
+        } else {
             return false;
         }
     }
 
     public static function responseSuccessWithData(
-        $http_code,$status_code,$api_code,$message_id,$message_en,$res_data
-    ){
+        $http_code, $status_code, $api_code, $message_id, $message_en, $res_data
+    )
+    {
         $response = [
             'http_response' => $http_code,
             'status_code' => $status_code,
             'api_code' => $api_code,
-            'message_id' =>  $message_id,
+            'message_id' => $message_id,
             'message_en' => $message_en,
             'res_data' => $res_data,
         ];
 
-        return response($response,$http_code);
+        return response($response, $http_code);
     }
 
     public static function responseErrorWithData(
-        $http_code,$status_code,$api_code,$message_id,$message_en,$res_data
-    ){
+        $http_code, $status_code, $api_code, $message_id, $message_en, $res_data
+    )
+    {
         $response = [
             'http_response' => $http_code,
             'status_code' => $status_code,
             'api_code' => $api_code,
-            'message_id' =>  $message_id,
+            'message_id' => $message_id,
             'message_en' => $message_en,
             'res_data' => $res_data,
         ];
 
-        return response($response,$http_code);
+        return response($response, $http_code);
     }
 
-    public static function removeFile($file_path){
+    public static function removeFile($file_path)
+    {
         // remove photo first
-        File::delete($file_path);
+        if (!Str::contains($file_path, 'razky_samples'))
+            File::delete($file_path);
         // if (file_exists($file_path)) {
         //     try {
         //         unlink($file_path);
@@ -81,6 +88,19 @@ class RazkyFeb
         return null;
     }
 
+    public static function insertNotification(
+        $userId, $title, $message, $desc, $type
+    )
+    {
+        $object = new UserMNotification();
+        $object->user_id = $userId;
+        $object->message = $message;
+        $object->title = $title;
+        $object->desc = $desc;
+        $object->type = $type;
+        $object->save();
+    }
+
     public static function profileImgPath()
     {
         return "photo/profile";
@@ -90,6 +110,7 @@ class RazkyFeb
     {
         return "photo/news";
     }
+
     public static function reqSellImgPath()
     {
         return "photo/profile";
